Module lock_keeper_key_server::database

Database for key-servers.

This database will hold information on users and the secret material they have stored in the key server.

Structs

SecretFilter

Filters that can be used to influence database queries. Any new database filters for secrets (e.g. created_time) should be added to this struct as optional fields. This will allow us to add new filters with minimal breakage.

Traits

DataStore

Defines the expected interface between a key server and its database.