Trait lock_keeper_key_server::database::DataStore

pub trait DataStore: Send + Sync + 'static {
    type Error: Error + Send + Sync + 'static;

    fn create_audit_event<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        actor: &'life1 AccountName,
        secret_id: &'life2 Option<KeyId>,
        action: ClientAction,
        status: EventStatus
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn find_audit_events<'life0, 'life1, 'async_trait>(
        &'life0 self,
        account_name: &'life1 AccountName,
        event_type: EventType,
        options: AuditEventOptions
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AuditEvent>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn add_user_secret<'life0, 'life1, 'async_trait>(
        &'life0 self,
        user_id: &'life1 UserId,
        secret: StoredSecret
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_user_secret<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        user_id: &'life1 UserId,
        key_id: &'life2 KeyId,
        filter: SecretFilter
    ) -> Pin<Box<dyn Future<Output = Result<StoredSecret, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn create_user<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        user_id: &'life1 UserId,
        account_name: &'life2 AccountName,
        server_registration: &'life3 ServerRegistration<OpaqueCipherSuite>
    ) -> Pin<Box<dyn Future<Output = Result<User, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait;
    fn find_user<'life0, 'life1, 'async_trait>(
        &'life0 self,
        account_name: &'life1 AccountName
    ) -> Pin<Box<dyn Future<Output = Result<Option<User>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn find_user_by_id<'life0, 'life1, 'async_trait>(
        &'life0 self,
        user_id: &'life1 UserId
    ) -> Pin<Box<dyn Future<Output = Result<Option<User>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn delete_user<'life0, 'life1, 'async_trait>(
        &'life0 self,
        user_id: &'life1 UserId
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn set_storage_key<'life0, 'life1, 'async_trait>(
        &'life0 self,
        user_id: &'life1 UserId,
        storage_key: Encrypted<StorageKey>
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

Defines the expected interface between a key server and its database.

This trait definition is not complete. New trait methods may be added in future server versions.

Required Associated Types

type Error: Error + Send + Sync + 'static

Required Methods

fn create_audit_event<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
actor: &'life1 AccountName,
secret_id: &'life2 Option<KeyId>,
action: ClientAction,
status: EventStatus
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Create a new [AuditEvent] for the given actor, action, and outcome

fn find_audit_events<'life0, 'life1, 'async_trait>(
&'life0 self,
account_name: &'life1 AccountName,
event_type: EventType,
options: AuditEventOptions
) -> Pin<Box<dyn Future<Output = Result<Vec<AuditEvent>, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Find [AuditEvent]s that correspond to the event type and provided filters

fn add_user_secret<'life0, 'life1, 'async_trait>(
&'life0 self,
user_id: &'life1 UserId,
secret: StoredSecret
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Add a [StoredSecret] to a [User]’s list of arbitrary secrets

fn get_user_secret<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
user_id: &'life1 UserId,
key_id: &'life2 KeyId,
filter: SecretFilter
) -> Pin<Box<dyn Future<Output = Result<StoredSecret, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Get a [User]’s [StoredSecret] based on its [KeyId]. A [StoredSecret] will only be returned if it matches the given SecretFilter.

fn create_user<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
user_id: &'life1 UserId,
account_name: &'life2 AccountName,
server_registration: &'life3 ServerRegistration<OpaqueCipherSuite>
) -> Pin<Box<dyn Future<Output = Result<User, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

Create a new [User] with their authentication information and insert it into the database.

fn find_user<'life0, 'life1, 'async_trait>(
&'life0 self,
account_name: &'life1 AccountName
) -> Pin<Box<dyn Future<Output = Result<Option<User>, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Find a [User] by their human-readable [AccountName].

fn find_user_by_id<'life0, 'life1, 'async_trait>(
&'life0 self,
user_id: &'life1 UserId
) -> Pin<Box<dyn Future<Output = Result<Option<User>, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Find a [User] by their machine-readable [UserId].

fn delete_user<'life0, 'life1, 'async_trait>(
&'life0 self,
user_id: &'life1 UserId
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Delete a [User] by their [UserId]

fn set_storage_key<'life0, 'life1, 'async_trait>(
&'life0 self,
user_id: &'life1 UserId,
storage_key: Encrypted<StorageKey>
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Set the storage_key field for the [User] associated with a given [UserId] Returns a LockKeeperServerError::InvalidAccount if the given user_id does not exist.