Enum lock_keeper_key_server::error::LockKeeperServerError

pub enum LockKeeperServerError {

    MissingService,
    PrivateKeyMissing,
    RemoteStorageKeyMissing,
    OpaqueServerSetupNotDefined,
    SessionCache(SessionCacheError),
    InvalidOpaqueDirectory,
    InvalidLogFilePath(PathBuf),
    AccountAlreadyRegistered,
    InvalidAccount,
    BlobSizeTooLarge,
    StorageKeyAlreadySet,
    StorageKeyNotSet,
    KeyNotFound,
    SessionIdNotFound,
    Bincode(Error),
    Database(DatabaseError),
    EnvVar(VarError),
    Hyper(Error),
    Io(Error),
    FileIo(Error, PathBuf),
    LockKeeper(LockKeeperError),
    OpaqueProtocol(ProtocolError),
    Rustls(Error),
    StrumParseError(ParseError),
    Toml(Error),
    TonicStatus(Status),
    TonicTransport(Error),
    WebPki(Error),

}

Variants

MissingService

PrivateKeyMissing

RemoteStorageKeyMissing

OpaqueServerSetupNotDefined

SessionCache(SessionCacheError)

InvalidOpaqueDirectory

InvalidLogFilePath(PathBuf)

AccountAlreadyRegistered

InvalidAccount

BlobSizeTooLarge

StorageKeyAlreadySet

StorageKeyNotSet

KeyNotFound

SessionIdNotFound

Bincode(Error)

Database(DatabaseError)

EnvVar(VarError)

Hyper(Error)

Io(Error)

Generic kitchen sink IO error. Use LockKeeperServerError::FileIo if the IO error is specifically related to working with files.

FileIo(Error, PathBuf)

IO error specific to file IO failing. Allows us to include the file that failed as part of the error.

LockKeeper(LockKeeperError)

OpaqueProtocol(ProtocolError)

Rustls(Error)

StrumParseError(ParseError)

Toml(Error)

TonicStatus(Status)

TonicTransport(Error)

WebPki(Error)