Module lock_keeper_key_server::operations::authenticate

Structs

Authenticate

AuthenticateStartResult ğŸ”’

Functions

authenticate_finish ğŸ”’

authenticate_start ğŸ”’

Returns the server-side start message along with a login result that will be used in the finish step.

send_user_id ğŸ”’

Returns the server-side start message along with a login result that will be used in the finish step.