Struct lock_keeper_key_server::server::channel::Authenticated

pub struct Authenticated<RNG: CryptoRng + RngCore> {
    pub session_key: OpaqueSessionKey,
    pub account: Account,
    pub rng: Arc<Mutex<RNG>>,
}

Passed to channel type as the AUTH generic parameter.

It is used for channels that handle authenticated operations. This type ensures that messages moving across a channel are encrypted.

Fields
session_key: OpaqueSessionKeyaccount: Accountrng: Arc<Mutex<RNG>>