Struct lock_keeper_key_server::server::channel::Channel

pub struct Channel<AUTH> {
    sender: Sender<Result<Message, Status>>,
    receiver: Streaming<Message>,
    metadata: RequestMetadata,
    auth: AUTH,
}

Server-side implementation of a two-way channel between a client and the server used to communicate Message objects.

This bidirectional Channel is implemented under the hood as two unidirectional channels. It can only be used with streaming messages generated by tonic.

These are tokio MPSC channels, but tonic is able to use them to send messages via the network between the server and client.

The AUTH generic parameter allows our channel to be generic over either authenticated or unauthenticated operations.

Fields
sender: Sender<Result<Message, Status>>

Sender end of an unidirectional channel. Allows us to send messages to the client. The server spawns the sender and receiver ends as a pair. The receiver is sent to the client as part of gRPC call so the server can send the client messages using this channel.

receiver: Streaming<Message>

A receiver end of an unidirectional channel. Allows us to receive messages from the client. When the client made a gRPC call, it sent this receiving end. The client can send messages to it and we will receive them.

metadata: RequestMetadataauth: AUTH
Implementations

impl<AUTH> Channel<AUTH>

pub fn metadata(&self) -> &RequestMetadata

Returns the metadata associated with this channel.

pub async fn send_error( &mut self, status: impl Into<Status> ) -> Result<(), LockKeeperError>

Send an error message across the channel.

pub async fn closed(&mut self)

impl<RNG: CryptoRng + RngCore> Channel<Authenticated<RNG>>

pub fn account_id(&self) -> AccountId

Returns the account ID for the authenticated user.

pub fn user_id(&self) -> &UserId

pub fn account(&self) -> &Account

Returns the full account info for the authenticated user.

pub fn set_storage_key(&mut self, storage_key: Encrypted<StorageKey>)

pub async fn receive<R: ConvertMessage>(&mut self) -> Result<R, LockKeeperError>

Receive the next message on the channel and convert it to the type R.

pub async fn send( &mut self, message: impl ConvertMessage ) -> Result<(), LockKeeperError>

Send a message across the channel. This function accepts any type that can be converted to a message via the [ConvertMessage] trait.

impl Channel<Unauthenticated>

pub fn new( request: Request<Streaming<Message>> ) -> Result<(Self, Receiver<Result<Message, Status>>), LockKeeperError>

pub fn into_authenticated<RNG: CryptoRng + RngCore>( self, account: Account, session_key: OpaqueSessionKey, rng: Arc<Mutex<RNG>> ) -> Channel<Authenticated<RNG>>

pub async fn receive<R: ConvertMessage>(&mut self) -> Result<R, LockKeeperError>

Receive the next message on the channel and convert it to the type R.

pub async fn send( &mut self, message: impl ConvertMessage ) -> Result<(), LockKeeperError>

Send a message across the channel. This function accepts any type that can be converted to a Message.