Trait lock_keeper_key_server::server::operation::Operation

pub(crate) trait Operation<AUTH: Send + 'static, DB: DataStore>: Sized + Send + 'static {
    fn operation<'life0, 'life1, 'async_trait>(
        self,
        channel: &'life0 mut ServerChannel<AUTH>,
        context: &'life1 mut Context<DB>
    ) -> Pin<Box<dyn Future<Output = Result<(), LockKeeperServerError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn handle_request<'async_trait>(
        self,
        context: Context<DB>,
        channel: ServerChannel<AUTH>
    ) -> Pin<Box<dyn Future<Output = Result<(), LockKeeperServerError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
    { ... }
}

A type implementing Operation can process tonic requests using a message-passing protocol facilitated by a ServerChannel.

Required Methods

fn operation<'life0, 'life1, 'async_trait>(
self,
channel: &'life0 mut ServerChannel<AUTH>,
context: &'life1 mut Context<DB>
) -> Pin<Box<dyn Future<Output = Result<(), LockKeeperServerError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Core logic for a given operation.

Provided Methods

fn handle_request<'async_trait>(
self,
context: Context<DB>,
channel: ServerChannel<AUTH>
) -> Pin<Box<dyn Future<Output = Result<(), LockKeeperServerError>> + Send + 'async_trait>>where
Self: 'async_trait,

Takes a request from tonic and spawns a new thread to process that request through the logic defined by the Operation::operation method. Any errors returned by the operation are logged and an appropriate error message is sent to the client.