Trait lock_keeper_key_server::server::session_cache::SessionCache

pub trait SessionCache: Send + Sync {
    fn create_session<'life0, 'async_trait>(
        &'life0 self,
        user_id: UserId,
        session_key: Encrypted<OpaqueSessionKey>
    ) -> Pin<Box<dyn Future<Output = Result<Uuid, SessionCacheError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn find_session<'life0, 'async_trait>(
        &'life0 self,
        session_id: Uuid,
        user_id: UserId
    ) -> Pin<Box<dyn Future<Output = Result<Session, SessionCacheError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn delete_session<'life0, 'async_trait>(
        &'life0 self,
        session_id: Uuid
    ) -> Pin<Box<dyn Future<Output = Result<(), SessionCacheError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Cache holding our sessions, per user, after authentication. Maps [UserId]s to Sessions. Sessions should be tagged with a timestamp. A session is considered invalid after the expiration time has elapsed.

Required Methods

fn create_session<'life0, 'async_trait>(
&'life0 self,
user_id: UserId,
session_key: Encrypted<OpaqueSessionKey>
) -> Pin<Box<dyn Future<Output = Result<Uuid, SessionCacheError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Store a newly created session for the specified user. The previous session for that user should be overwritten.

fn find_session<'life0, 'async_trait>(
&'life0 self,
session_id: Uuid,
user_id: UserId
) -> Pin<Box<dyn Future<Output = Result<Session, SessionCacheError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Get the session for the specified user, if one exists. This function should check if the session has expired and return an error if so.

fn delete_session<'life0, 'async_trait>(
&'life0 self,
session_id: Uuid
) -> Pin<Box<dyn Future<Output = Result<(), SessionCacheError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Indicate that the session for this user has expired. If the same user attempts to make a server call after expiring their session, they should need to authenticate again first.