Trait lock_keeper_key_server::server::session_key_cache::SessionCache

pub trait SessionCache: Send + Sync {
    fn create_session(&mut self, user: UserId, key: OpaqueSessionKey);
    fn find_session(
        &mut self,
        user: UserId
    ) -> Result<OpaqueSessionKey, SessionCacheError>;
    fn delete_session(
        &mut self,
        user_id: UserId
    ) -> Result<(), SessionCacheError>;
}

Cache holding our session keys, per user, after authentication. Maps [UserId]s to [OpaqueSessionKey]s. Keys should be tagged with a timestamp. A key is considered invalid after the expiration time has elapsed.

Required Methods

fn create_session(&mut self, user: UserId, key: OpaqueSessionKey)

Add a new session key for the specified user. The previous key for that user should be overwritten.

fn find_session(
&mut self,
user: UserId
) -> Result<OpaqueSessionKey, SessionCacheError>

Get the session key for the specified user, if one exists. This function should check if the key has expired and return an error if so.

fn delete_session(&mut self, user_id: UserId) -> Result<(), SessionCacheError>

Indicate that the session key for this user has expired. If the same user attempts to make a server call after expiring their session key, they should need to authenticate again first.