Module lock_keeper_key_server::operations

Modules

authenticate ğŸ”’

create_storage_key ğŸ”’

generate_secret ğŸ”’

import_signing_key ğŸ”’

Import signing key operation. This protocol allows the client to import a key into the server.

logout ğŸ”’

register ğŸ”’

remote_generate_signing_key ğŸ”’

Client has asked server to generate a new signing key and store it in the server.

remote_sign_bytes ğŸ”’

This operation allows client to specify a key ID for a key that was remotely generated on the server and use this key to sign a message.

retrieve_audit_events ğŸ”’

This operation allows client to retrieve the stored audit event logs.

retrieve_secret ğŸ”’

retrieve_storage_key ğŸ”’

Structs

Authenticate

CreateStorageKey

GenerateSecret

ImportSigningKey

Logout

Register

RemoteGenerateSigningKey

RemoteSignBytes

RetrieveAuditEvents

RetrieveSecret

RetrieveStorageKey