Trait lock_keeper_key_server::server::database::DataStore

pub trait DataStore: Send + Sync + 'static {
    fn create_audit_event<'life0, 'life1, 'async_trait>(
        &'life0 self,
        request_id: Uuid,
        account_id: AccountId,
        key_id: &'life1 Option<KeyId>,
        action: ClientAction,
        status: EventStatus
    ) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn find_audit_events<'life0, 'async_trait>(
        &'life0 self,
        account_id: AccountId,
        event_type: EventType,
        options: AuditEventOptions
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AuditEvent>, DatabaseError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn add_secret<'life0, 'async_trait>(
        &'life0 self,
        secret: StoredSecret
    ) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_secret<'life0, 'life1, 'async_trait>(
        &'life0 self,
        account_id: AccountId,
        key_id: &'life1 KeyId,
        filter: SecretFilter
    ) -> Pin<Box<dyn Future<Output = Result<StoredSecret, DatabaseError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn delete_secret<'life0, 'life1, 'async_trait>(
        &'life0 self,
        account_id: AccountId,
        key_id: &'life1 KeyId
    ) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn create_account<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        user_id: &'life1 UserId,
        account_name: &'life2 AccountName,
        server_registration: &'life3 ServerRegistration<OpaqueCipherSuite>
    ) -> Pin<Box<dyn Future<Output = Result<Account, DatabaseError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait;
    fn find_account_by_name<'life0, 'life1, 'async_trait>(
        &'life0 self,
        account_name: &'life1 AccountName
    ) -> Pin<Box<dyn Future<Output = Result<Option<Account>, DatabaseError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn find_account<'life0, 'async_trait>(
        &'life0 self,
        account_id: AccountId
    ) -> Pin<Box<dyn Future<Output = Result<Option<Account>, DatabaseError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn delete_account<'life0, 'async_trait>(
        &'life0 self,
        account_id: AccountId
    ) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn set_storage_key<'life0, 'async_trait>(
        &'life0 self,
        account_id: AccountId,
        storage_key: Encrypted<StorageKey>
    ) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn user_id_exists<'life0, 'life1, 'async_trait>(
        &'life0 self,
        user_id: &'life1 UserId
    ) -> Pin<Box<dyn Future<Output = Result<bool, DatabaseError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
}

Defines the expected interface between a key server and its database.

This trait definition is not complete. New trait methods may be added in future server versions.

Required Methods

fn create_audit_event<'life0, 'life1, 'async_trait>(
&'life0 self,
request_id: Uuid,
account_id: AccountId,
key_id: &'life1 Option<KeyId>,
action: ClientAction,
status: EventStatus
) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Create a new AuditEvent for the given actor, action, and outcome

fn find_audit_events<'life0, 'async_trait>(
&'life0 self,
account_id: AccountId,
event_type: EventType,
options: AuditEventOptions
) -> Pin<Box<dyn Future<Output = Result<Vec<AuditEvent>, DatabaseError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Find AuditEvents that correspond to the event type and provided filters

fn add_secret<'life0, 'async_trait>(
&'life0 self,
secret: StoredSecret
) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Add a StoredSecret to a Account’s list of arbitrary secrets

fn get_secret<'life0, 'life1, 'async_trait>(
&'life0 self,
account_id: AccountId,
key_id: &'life1 KeyId,
filter: SecretFilter
) -> Pin<Box<dyn Future<Output = Result<StoredSecret, DatabaseError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Get a Account’s StoredSecret based on its KeyId. A StoredSecret will only be returned if it matches the given SecretFilter.

fn delete_secret<'life0, 'life1, 'async_trait>(
&'life0 self,
account_id: AccountId,
key_id: &'life1 KeyId
) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Delete a StoredSecret from an Account’s list of secrets.

fn create_account<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
user_id: &'life1 UserId,
account_name: &'life2 AccountName,
server_registration: &'life3 ServerRegistration<OpaqueCipherSuite>
) -> Pin<Box<dyn Future<Output = Result<Account, DatabaseError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,

Create a new Account with their authentication information and insert it into the database.

fn find_account_by_name<'life0, 'life1, 'async_trait>(
&'life0 self,
account_name: &'life1 AccountName
) -> Pin<Box<dyn Future<Output = Result<Option<Account>, DatabaseError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Find a Account by their human-readable AccountName.

fn find_account<'life0, 'async_trait>(
&'life0 self,
account_id: AccountId
) -> Pin<Box<dyn Future<Output = Result<Option<Account>, DatabaseError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Find a Account by their AccountId.

fn delete_account<'life0, 'async_trait>(
&'life0 self,
account_id: AccountId
) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Delete a Account by their AccountId

fn set_storage_key<'life0, 'async_trait>(
&'life0 self,
account_id: AccountId,
storage_key: Encrypted<StorageKey>
) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Set the storage_key field for the Account associated with a given AccountId Returns a LockKeeperServerError::InvalidAccount if the given account_id does not exist.

fn user_id_exists<'life0, 'life1, 'async_trait>(
&'life0 self,
user_id: &'life1 UserId
) -> Pin<Box<dyn Future<Output = Result<bool, DatabaseError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Returns true if the UserId already exists in the database.